Fixed functional tests
